+++ NEWS.md
@@ -1,3 +1,12 @@
+# bcdata 0.5.1
+
+* Fix bugs where `bcdc_search_facets()`, `bcdc_list_group_records()` and 
+  `bcdc_list_organization_records()` were not returning all records after 
+  the upgrade to CKAN 2.9 (#353). Includes adding tests and bcdata-specific 
+  options for these queries, see `bcdc_options()`.
+* Added examples of partial query using `%like%` and `%in%` in 
+  `bcdc_query_geodata()` (#356, thanks @bevingtona).
+
 # bcdata 0.5.0
 
 * Make functions more robust to non-functioning WMS/WFS GetCapabilities requests